refactor: reorganize app structure and packages
