+++ src/app/crm/page.tsx
@@ -1,0 +1,7 @@
+export default function CRMPage() {
+  return (
+    <div className="p-4">
+      <h1>CRM</h1>
+    </div>
+  );
+}
