+++ src/app/dashboard/sections/index.ts
@@ -1,0 +1,4 @@
+export { default as ContactsSection } from "./contacts";
+export { default as EmailSection } from "./email";
+export { default as FinanceSection } from "./finance";
+export { default as SocialSection } from "./social";
