+++ src/app/marketing/page.tsx
@@ -1,0 +1,7 @@
+export default function MarketingPage() {
+  return (
+    <div className="p-4">
+      <h1>Marketing</h1>
+    </div>
+  );
+}
