+++ src/packages/db/server.ts
@@ -1,0 +1,2 @@
+// Placeholder for server-side Supabase utilities
+export {};
