+++ src/packages/emailer/README.md
@@ -1,0 +1,3 @@
+# Emailer Package
+
+Email templating utilities placeholder.
