+++ src/packages/finance/README.md
@@ -1,0 +1,3 @@
+# Finance Package
+
+Finance SDK wrapper placeholder.
