+++ src/packages/ui/README.md
@@ -1,0 +1,3 @@
+# UI Package
+
+Shared ShadCN components placeholder.
