+++ src/tests/components/README.md
@@ -1,0 +1,1 @@
+Placeholder for component tests.
