+++ src/tests/pages/README.md
@@ -1,0 +1,1 @@
+Placeholder for page tests.
